backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
sqlalchemy==2.0.7
pydantic==1.10.7
psycopg2-binary==2.9.5
alembic==1.10.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode==7.4.2
pillow==9.5.0
pytest==7.3.1
httpx==0.24.0 